move creation of ColumnarBatch to InMemoryTableScanExec.inputRDDs
